requirements.txt
attrs==19.3.0
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==8.2.0
mysql-connector-python==8.0.20
numpy==1.18.4
packaging==20.3
pandas==1.0.3
pluggy==0.13.1
protobuf==3.12.0
py==1.8.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.2
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
SQLAlchemy==1.3.17
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
